backend/bot/requirements.txt
aiogram>=3.0
vkbottle>=4.3
python-dotenv>=1.0.0
firebase-admin>=6.0.0
PyJWT>=2.8.0
aiohttp>=3.10.0
